las tareas se eliminan dinámicamente

+++ src/pages/ManageAccount.jsx
@@ -6,10 +6,12 @@ import ModalLibro from "../components/ModalLibro";
 import { useEffect } from "react";
 import { getDataUser } from "../services/fetch";
 import { deleteData } from "../services/fetch";
+import { acceptPopUp } from "../services/alertas";
 
 const ManageAccount = () => { //CRUD
     const [modalShow, setModalShow] = useState(false);
     const [libros, setLibros] = useState([])
+    const [recargar, setRecargar] = useState(false)
 
     useEffect(()=>{
         const librosUser = async ()=>{
@@ -17,19 +19,26 @@ const ManageAccount = () => { //CRUD
             setLibros(dataLibroUser)
         }
         librosUser()
-    },[])
+    },[libros])
     
-        const deleteLibro = async(id) =>{
-            await deleteData("libros",id)
-            
+    const deleteLibro = async(id) =>{
+        const isConfirmed = await acceptPopUp("Estás intentando eliminar un libro, ¿Continuar?", "El libro se eliminó con éxito", "La eliminación del libro fue cancelada");
+   if (isConfirmed) {
+            await deleteData("libros", id);
         }
-    
+
+    }
+    //DE MOMENTO
+    const recargaPag =()=>{
+        setRecargar(!recargar)
+    }
+
     return (
         <>
             <Navbar />
             <h1>My account</h1>
 
-            <ListaLibros cardLibro={libros} btnEditar={""} btnEliminar={deleteLibro()} />
+            <ListaLibros cardLibro={libros} btnEditar={""} btnEliminar={deleteLibro} />
 
             <Button variant="primary" onClick={() => setModalShow(true)}>
                 Add a new book
@@ -37,8 +46,8 @@ const ManageAccount = () => { //CRUD
             <ModalLibro
                 show={modalShow}
                 onHide={() => setModalShow(false)}
+                recargar={recargaPag}
             />
-            
         </>
     )
 }
